Add UUID type checks to bucket controller

+++ src/Controller/BucketsController.php
@@ -14,8 +14,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use App\Models\User;
 use App\Models\Apikeys;
 use App\Models\Buckets;
-
-//use App\Utils\Sentry;
+use App\Utils\Auth;
 
 class BucketsController extends AbstractController {
 
@@ -26,16 +25,26 @@ class BucketsController extends AbstractController {
      */
 
 	public function create_user_bucket(Request $request, $id){
+		$auth = new Auth();
 		$buckets = new Buckets();
 
-		if($request->request->has('bucket_name') && $request->request->has('password')){
-			$new_bucket = $buckets->create_new_user_bucket($id, $request->request->get('bucket_name'), $request->request->get('password'));
+		if($auth->isValidUUID($id)){
 
-			return new Response(json_encode($new_bucket));
+			if($request->request->has('bucket_name') && $request->request->has('password')){
+				$new_bucket = $buckets->create_new_user_bucket($id, $request->request->get('bucket_name'), $request->request->get('password'));
+
+				return new Response(json_encode($new_bucket));
+		
+			}else{
+
+				return new Response(json_encode(array('success' => false, 'errorcode' => 302882)));
+		
+			}
 		}else{
-			return new Response(json_encode(array('success' => false, 'errorcode' => 302882)));
+
+			return new Response(json_encode(array('success' => false, 'errorcode' => 302883)));
+
 		}
-	}
 
 	/**
      * Matches /buckets/{id}/delete exactly
@@ -44,14 +53,25 @@ class BucketsController extends AbstractController {
      */
 
 	public function delete_user_bucket(Request $request, $id){
+		$auth = new Auth();
 		$buckets = new Buckets();
 
-		if($request->request->has('bucket_name') && $request->request->has('password')){
-			$deleted_bucket = $buckets->delete_user_bucket($id, $request->request->get('bucket_name'), $request->request->get('password'));
+		if($auth->isValidUUID($id)){
+
+			if($request->request->has('bucket_name') && $request->request->has('password')){
+				$deleted_bucket = $buckets->delete_user_bucket($id, $request->request->get('bucket_name'), $request->request->get('password'));
 
 			return new Response(json_encode($deleted_bucket));
+		
+			}else{
+		
+				return new Response(json_encode(array('success' => false, 'errorcode' => 302882)));
+		
+			}
 		}else{
-			return new Response(json_encode(array('success' => false, 'errorcode' => 302882)));
+
+			return new Response(json_encode(array('success' => false, 'errorcode' => 302883)));
+
 		}
 	}
 
@@ -62,14 +82,24 @@ class BucketsController extends AbstractController {
      */
 
 	public function assign_domain_to_bucket(Request $request, $id){
+		$auth = new Auth();
 		$buckets = new Buckets();
 
-		if($request->request->has('bucket_name') && $request->request->has('password') && $request->request->has('domain')){
-			$assign_domain = $buckets->assign_domain_to_bucket($id, $request->request->get('password'), $request->request->get('bucket_name'), $request->request->get('domain'));
+		if($auth->isValidUUID($id)){
 
-			return new Response(json_encode($assign_domain));
+			if($request->request->has('bucket_name') && $request->request->has('password') && $request->request->has('domain')){
+				$assign_domain = $buckets->assign_domain_to_bucket($id, $request->request->get('password'), $request->request->get('bucket_name'), $request->request->get('domain'));
+
+				return new Response(json_encode($assign_domain));
+			
+			}else{
+		
+				return new Response(json_encode(array('success' => false, 'errorcode' => 302882)));
+			}
 		}else{
-			return new Response(json_encode(array('success' => false, 'errorcode' => 302882)));
+
+			return new Response(json_encode(array('success' => false, 'errorcode' => 302883)));
+
 		}
 	}
